divide Auth & User controller/service/testcase
