+++ src/main/java/com/example/mailService/controller/AuthController.java
@@ -1,47 +1,46 @@
 package com.example.mailService.controller;
 
+import com.example.mailService.domain.dto.JwtTokenDto;
 import com.example.mailService.domain.dto.UserDto;
 import com.example.mailService.domain.dto.UserLoginDto;
 import com.example.mailService.domain.dto.UserSignUpDto;
 import com.example.mailService.domain.entity.User;
+import com.example.mailService.service.AuthService;
 import com.example.mailService.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/auth")
 public class AuthController {
 
-    private final AuthenticationManager authenticationManager;
-
-    private final UserService userService;
+    private final AuthService authService;
 
     @PostMapping("/login")
     public ResponseEntity<?> userLogin(@Valid @RequestBody UserLoginDto loginDto) {
+        JwtTokenDto tokenDto = authService.loginUser(loginDto);
 
-        Authentication authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        loginDto.getUsername(),
-                        loginDto.getPassword()
-                )
-        );
+        if (Objects.isNull(tokenDto)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
 
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        // TODO: add tokenGeneration before response
-        return ResponseEntity.ok(null);
+        // Authorization header 에 JWT 토큰 설정
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "Bearer " + tokenDto.getToken());
+
+        return new ResponseEntity<>(tokenDto, headers, HttpStatus.OK);
     }
 
     @PostMapping("/signup")
     public UserDto userSignUp(@Valid @RequestBody UserSignUpDto signUpDto) {
-        User user = userService.registerUser(signUpDto);
+        User user = authService.registerUser(signUpDto);
         return UserDto.from(user);
     }
 }
