+++ src/main/java/com/example/mailService/domain/dto/JwtTokenDto.java
@@ -1,0 +1,11 @@
+package com.example.mailService.domain.dto;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class JwtTokenDto {
+    private String token;
+}
